requirements.txt
fastapi
uvicorn
python-dotenv
beautifulsoup4
openai
httpx
jinja2
